+++ framework/src/main/java/org/checkerframework/framework/type/GenericAnnotatedTypeFactory.java
@@ -1182,36 +1182,14 @@ public abstract class GenericAnnotatedTypeFactory<
   /**
    * Returns the store immediately before a given {@link Tree}.
    *
+   * @param tree the tree
    * @return the store immediately before a given {@link Tree}
    */
-  public Store getStoreBefore(Tree tree) {
-    if (!analysis.isRunning()) {
-      return flowResult.getStoreBefore(tree);
+  public @Nullable Store getStoreBefore(Tree tree) {
+    if (analysis.isRunning()) {
+      return analysis.getStoreBefore(tree, flowResultAnalysisCaches);
     }
-    Set<Node> nodes = analysis.getNodesForTree(tree);
-    if (nodes != null) {
-      return getStoreBefore(nodes);
-    } else {
-      return flowResult.getStoreBefore(tree);
-    }
-  }
-
-  /**
-   * Returns the store immediately before a given Set of {@link Node}s.
-   *
-   * @return the store immediately before a given Set of {@link Node}s
-   */
-  public Store getStoreBefore(Set<Node> nodes) {
-    Store merge = null;
-    for (Node aNode : nodes) {
-      Store s = getStoreBefore(aNode);
-      if (merge == null) {
-        merge = s;
-      } else if (s != null) {
-        merge = merge.leastUpperBound(s);
-      }
-    }
-    return merge;
+    return flowResult.getStoreBefore(tree);
   }
 
   /**
@@ -1221,21 +1199,10 @@ public abstract class GenericAnnotatedTypeFactory<
    * @return the store immediately before {@code node}
    */
   public @Nullable Store getStoreBefore(Node node) {
-    if (!analysis.isRunning()) {
-      return flowResult.getStoreBefore(node);
+    if (analysis.isRunning()) {
+      return analysis.getStoreBefore(node, flowResultAnalysisCaches);
     }
-    TransferInput<Value, Store> prevStore = analysis.getInput(node.getBlock());
-    if (prevStore == null) {
-      return null;
-    }
-    Store store =
-        AnalysisResult.runAnalysisFor(
-            node,
-            Analysis.BeforeOrAfter.BEFORE,
-            prevStore,
-            analysis.getNodeValues(),
-            flowResultAnalysisCaches);
-    return store;
+    return flowResult.getStoreBefore(node);
   }
 
   /**
@@ -1247,30 +1214,10 @@ public abstract class GenericAnnotatedTypeFactory<
    * @return the store immediately after a given tree
    */
   public @Nullable Store getStoreAfter(Tree tree) {
-    if (!analysis.isRunning()) {
-      return flowResult.getStoreAfter(tree);
+    if (analysis.isRunning()) {
+      return analysis.getStoreAfter(tree, flowResultAnalysisCaches);
     }
-    Set<Node> nodes = analysis.getNodesForTree(tree);
-    return getStoreAfter(nodes);
-  }
-
-  /**
-   * Returns the store immediately after a given set of nodes.
-   *
-   * @param nodes the nodes whose post-stores to LUB
-   * @return the LUB of the stores store immediately after {@code nodes}
-   */
-  public Store getStoreAfter(Set<Node> nodes) {
-    Store merge = null;
-    for (Node node : nodes) {
-      Store s = getStoreAfter(node);
-      if (merge == null) {
-        merge = s;
-      } else if (s != null) {
-        merge = merge.leastUpperBound(s);
-      }
-    }
-    return merge;
+    return flowResult.getStoreAfter(tree);
   }
 
   /**
@@ -1280,17 +1227,10 @@ public abstract class GenericAnnotatedTypeFactory<
    * @return the store immediately after a given {@link Node}
    */
   public Store getStoreAfter(Node node) {
-    if (!analysis.isRunning()) {
-      return flowResult.getStoreAfter(node);
+    if (analysis.isRunning()) {
+      return analysis.getStoreAfter(node, flowResultAnalysisCaches);
     }
-    Store res =
-        AnalysisResult.runAnalysisFor(
-            node,
-            Analysis.BeforeOrAfter.AFTER,
-            analysis.getInput(node.getBlock()),
-            analysis.getNodeValues(),
-            flowResultAnalysisCaches);
-    return res;
+    return flowResult.getStoreAfter(node);
   }
 
   /**